# [Enhancement] Add hashJoin execution cost model