--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/LogicalProperty.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/base/LogicalProperty.java
@@ -3,7 +3,6 @@
 package com.starrocks.sql.optimizer.base;
 
 import com.google.common.base.Preconditions;
-import com.starrocks.common.FeConstants;
 import com.starrocks.sql.optimizer.ExpressionContext;
 import com.starrocks.sql.optimizer.operator.Operator;
 import com.starrocks.sql.optimizer.operator.OperatorVisitor;
@@ -25,8 +24,7 @@ import com.starrocks.sql.optimizer.operator.logical.MockOperator;
 public class LogicalProperty implements Property {
     // Operator's output columns
     private ColumnRefSet outputColumns;
-    // The tablets num of left most scan node
-    private int leftMostScanTabletsNum;
+
     // The flag for execute upon less than or equal one tablet
     private boolean isExecuteInOneTablet;
 
@@ -38,16 +36,8 @@ public class LogicalProperty implements Property {
         this.outputColumns = outputColumns;
     }
 
-    public int getLeftMostScanTabletsNum() {
-        return leftMostScanTabletsNum;
-    }
-
     public boolean isExecuteInOneTablet() {
         return isExecuteInOneTablet;
-    }
-
-    public void setLeftMostScanTabletsNum(int leftMostScanTabletsNum) {
-        this.leftMostScanTabletsNum = leftMostScanTabletsNum;
     }
 
     public LogicalProperty() {
@@ -60,62 +50,13 @@ public class LogicalProperty implements Property {
 
     public LogicalProperty(LogicalProperty other) {
         outputColumns = other.outputColumns.clone();
-        leftMostScanTabletsNum = other.leftMostScanTabletsNum;
         isExecuteInOneTablet = other.isExecuteInOneTablet;
     }
 
     public void derive(ExpressionContext expressionContext) {
         LogicalOperator op = (LogicalOperator) expressionContext.getOp();
         outputColumns = op.getOutputColumns(expressionContext);
-        leftMostScanTabletsNum = op.accept(new LeftMostScanTabletsNumVisitor(), expressionContext);
         isExecuteInOneTablet = op.accept(new OneTabletExecutorVisitor(), expressionContext);
-    }
-
-    static class LeftMostScanTabletsNumVisitor extends OperatorVisitor<Integer, ExpressionContext> {
-        @Override
-        public Integer visitOperator(Operator node, ExpressionContext context) {
-            Preconditions.checkState(context.arity() != 0);
-            return context.getChildLeftMostScanTabletsNum(0);
-        }
-
-        @Override
-        public Integer visitMockOperator(MockOperator node, ExpressionContext context) {
-            return 1;
-        }
-
-        @Override
-        public Integer visitLogicalTableScan(LogicalScanOperator node, ExpressionContext context) {
-            if (node instanceof LogicalOlapScanOperator) {
-                return ((LogicalOlapScanOperator) node).getSelectedTabletId().size();
-            } else {
-                // It's very hard to estimate how many tablets scanned by this operator,
-                // because some operator even does not have the concept of tablets.
-                // The value should not be too low, otherwise it will make cost optimizer to underestimate the cost of broadcast.
-                // A thing to be noted that, this tablet number is better not to be 1, to avoid generate 1 phase agg.
-                return FeConstants.DEFAULT_TABLET_NUMBER;
-            }
-        }
-
-        @Override
-        public Integer visitLogicalValues(LogicalValuesOperator node, ExpressionContext context) {
-            return 1;
-        }
-
-        @Override
-        public Integer visitLogicalTableFunction(LogicalTableFunctionOperator node, ExpressionContext context) {
-            return 1;
-        }
-
-        @Override
-        public Integer visitLogicalCTEAnchor(LogicalCTEAnchorOperator node, ExpressionContext context) {
-            Preconditions.checkState(context.arity() == 2);
-            return context.getChildLeftMostScanTabletsNum(1);
-        }
-
-        @Override
-        public Integer visitLogicalCTEConsume(LogicalCTEConsumeOperator node, ExpressionContext context) {
-            return 2;
-        }
     }
 
     static class OneTabletExecutorVisitor extends OperatorVisitor<Boolean, ExpressionContext> {