--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/cost/CostModel.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/cost/CostModel.java
@@ -9,11 +9,13 @@ import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.sql.common.ErrorType;
 import com.starrocks.sql.common.StarRocksPlannerException;
 import com.starrocks.sql.optimizer.ExpressionContext;
+import com.starrocks.sql.optimizer.Group;
 import com.starrocks.sql.optimizer.GroupExpression;
 import com.starrocks.sql.optimizer.JoinHelper;
 import com.starrocks.sql.optimizer.Utils;
 import com.starrocks.sql.optimizer.base.ColumnRefSet;
 import com.starrocks.sql.optimizer.base.DistributionSpec;
+import com.starrocks.sql.optimizer.base.PhysicalPropertySet;
 import com.starrocks.sql.optimizer.operator.Operator;
 import com.starrocks.sql.optimizer.operator.OperatorType;
 import com.starrocks.sql.optimizer.operator.OperatorVisitor;
@@ -35,24 +37,52 @@ import com.starrocks.sql.optimizer.operator.physical.PhysicalWindowOperator;
 import com.starrocks.sql.optimizer.operator.scalar.BinaryPredicateOperator;
 import com.starrocks.sql.optimizer.statistics.Statistics;
 import com.starrocks.statistic.StatsConstants;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class CostModel {
+
+    private static final Logger LOG = LogManager.getLogger(CostModel.class);
     public static double calculateCost(GroupExpression expression) {
         ExpressionContext expressionContext = new ExpressionContext(expression);
         return calculateCost(expressionContext);
     }
 
     private static double calculateCost(ExpressionContext expressionContext) {
-        CostEstimator costEstimator = new CostEstimator();
+        CostEstimator costEstimator = new CostEstimator(null);
         CostEstimate costEstimate = expressionContext.getOp().accept(costEstimator, expressionContext);
-        return getRealCost(costEstimate);
+        double realCost = getRealCost(costEstimate);
+        LOG.debug("operator: {}, outputRowCount: {}, outPutSize: {}, costEstimate: {}, realCost: {}",
+                expressionContext.getOp(),
+                expressionContext.getStatistics().getOutputRowCount(),
+                expressionContext.getStatistics().getComputeSize(),
+                costEstimate, realCost);
+        return realCost;
     }
 
     public static CostEstimate calculateCostEstimate(ExpressionContext expressionContext) {
-        CostEstimator costEstimator = new CostEstimator();
+        CostEstimator costEstimator = new CostEstimator(null);
         return expressionContext.getOp().accept(costEstimator, expressionContext);
+    }
+
+    public static double calculateCostWithInputProperty(GroupExpression expression,
+                                                              List<PhysicalPropertySet> inputProperties) {
+        ExpressionContext expressionContext = new ExpressionContext(expression);
+        CostEstimator costEstimator = new CostEstimator(inputProperties);
+        CostEstimate costEstimate = expressionContext.getOp().accept(costEstimator, expressionContext);
+        double realCost = getRealCost(costEstimate);
+
+        LOG.debug("operator: {}, group id: {}, child group id: {}, " +
+                        "inputProperties: {}, outputRowCount: {}, outPutSize: {}, costEstimate: {}, realCost: {}",
+                expressionContext.getOp(), expression.getGroup().getId(),
+                expression.getInputs().stream().map(Group::getId).collect(Collectors.toList()), inputProperties,
+                expressionContext.getStatistics().getOutputRowCount(),
+                expressionContext.getStatistics().getComputeSize(),
+                costEstimate, realCost);
+        return realCost;
     }
 
     public static double getRealCost(CostEstimate costEstimate) {
@@ -72,6 +102,13 @@ public class CostModel {
     }
 
     private static class CostEstimator extends OperatorVisitor<CostEstimate, ExpressionContext> {
+
+        private final List<PhysicalPropertySet> inputProperties;
+
+        private CostEstimator(List<PhysicalPropertySet> inputProperties) {
+            this.inputProperties = inputProperties;
+        }
+
         @Override
         public CostEstimate visitOperator(Operator node, ExpressionContext context) {
             return CostEstimate.zero();
@@ -195,11 +232,12 @@ public class CostModel {
             ConnectContext ctx = ConnectContext.get();
             SessionVariable sessionVariable = ctx.getSessionVariable();
             DistributionSpec distributionSpec = node.getDistributionSpec();
+            double outputSize = statistics.getOutputSize(outputColumns);
             // set network start cost 1 at least
             // avoid choose network plan when the cost is same as colocate plans
             switch (distributionSpec.getType()) {
                 case ANY:
-                    result = CostEstimate.ofCpu(statistics.getOutputSize(outputColumns));
+                    result = CostEstimate.ofCpu(outputSize);
                     break;
                 case BROADCAST:
                     int parallelExecInstanceNum = getParallelExecInstanceNum(
@@ -207,14 +245,15 @@ public class CostModel {
                     // beNum is the number of right table should broadcast, now use alive backends
                     int aliveBackendNumber = ctx.getAliveBackendNumber();
                     int beNum = Math.max(1, aliveBackendNumber);
-                    result = CostEstimate.of(statistics.getOutputSize(outputColumns) * aliveBackendNumber,
-                            statistics.getOutputSize(outputColumns) * beNum * parallelExecInstanceNum,
-                            Math.max(statistics.getOutputSize(outputColumns) * beNum * parallelExecInstanceNum, 1));
-                    if (statistics.getOutputSize(outputColumns) > sessionVariable.getMaxExecMemByte()) {
-                        return CostEstimate.of(result.getCpuCost() * StatsConstants.BROADCAST_JOIN_MEM_EXCEED_PENALTY,
-                                result.getMemoryCost() * StatsConstants.BROADCAST_JOIN_MEM_EXCEED_PENALTY,
-                                result.getNetworkCost() * StatsConstants.BROADCAST_JOIN_MEM_EXCEED_PENALTY);
+
+                    result = CostEstimate.of(outputSize * aliveBackendNumber,
+                            outputSize * beNum * parallelExecInstanceNum,
+                            Math.max(outputSize * beNum * parallelExecInstanceNum, 1));
+                    if (outputSize > sessionVariable.getMaxExecMemByte()) {
+                        result = result.multiplyBy(StatsConstants.BROADCAST_JOIN_MEM_EXCEED_PENALTY);
                     }
+                    LOG.debug("parallelInstNum: {}, beNum: {}, aliveBeNum: {}, outputSize: {}.",
+                            parallelExecInstanceNum, aliveBackendNumber, beNum, outputSize);
                     break;
                 case SHUFFLE:
                     // This is used to generate "ScanNode->LocalShuffle->OnePhaseLocalAgg" for the single backend,
@@ -224,12 +263,12 @@ public class CostModel {
                     boolean ignoreNetworkCost = sessionVariable.isEnableLocalShuffleAgg()
                             && sessionVariable.isEnablePipelineEngine()
                             && GlobalStateMgr.getCurrentSystemInfo().isSingleBackendAndComputeNode();
-                    double networkCost = ignoreNetworkCost ? 0 : Math.max(statistics.getOutputSize(outputColumns), 1);
+                    double networkCost = ignoreNetworkCost ? 0 : Math.max(outputSize, 1);
 
-                    result = CostEstimate.of(statistics.getOutputSize(outputColumns), 0, networkCost);
+                    result = CostEstimate.of(outputSize, 0, networkCost);
                     break;
                 case GATHER:
-                    result = CostEstimate.of(statistics.getOutputSize(outputColumns), 0,
+                    result = CostEstimate.of(outputSize, 0,
                             Math.max(statistics.getOutputSize(outputColumns), 1));
                     break;
                 default:
@@ -237,16 +276,13 @@ public class CostModel {
                             "not support " + distributionSpec.getType() + "distribution type",
                             ErrorType.UNSUPPORTED);
             }
+            LOG.debug("distribution type {}, cost: {}.", distributionSpec.getType(), result);
             return result;
         }
 
         @Override
         public CostEstimate visitPhysicalHashJoin(PhysicalHashJoinOperator join, ExpressionContext context) {
             Preconditions.checkState(context.arity() == 2);
-            // For broadcast join, use leftExecInstanceNum as right child real destinations num.
-            int leftExecInstanceNum = context.getChildLeftMostScanTabletsNum(0);
-            context.getChildLogicalProperty(1).setLeftMostScanTabletsNum(leftExecInstanceNum);
-
             Statistics statistics = context.getStatistics();
             Preconditions.checkNotNull(statistics);
 
@@ -260,10 +296,8 @@ public class CostModel {
 
             Preconditions.checkState(!(join.getJoinType().isCrossJoin() || eqOnPredicates.isEmpty()),
                     "should be handled by nestloopjoin");
-
-            return CostEstimate.of(leftStatistics.getOutputSize(context.getChildOutputColumns(0))
-                            + rightStatistics.getOutputSize(context.getChildOutputColumns(1)),
-                    rightStatistics.getOutputSize(context.getChildOutputColumns(1)), 0);
+            HashJoinCostModel joinCostModel = new HashJoinCostModel(context, inputProperties, eqOnPredicates);
+            return CostEstimate.of(joinCostModel.getCpuCost(), joinCostModel.getMemCost(), 0);
         }
 
         @Override