--- a/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalDistributionOperator.java
+++ b/fe/fe-core/src/main/java/com/starrocks/sql/optimizer/operator/physical/PhysicalDistributionOperator.java
@@ -48,4 +48,9 @@ public class PhysicalDistributionOperator extends PhysicalOperator {
     public boolean couldApplyStringDict(Set<Integer> childDictColumns) {
         return true;
     }
+
+    @Override
+    public String toString() {
+        return String.format("[optType: %s, distributionSpec: %s]", opType, distributionSpec);
+    }
 }