… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/cost/HashJoinCostModel.java
… fe/fe-core/src/main/java/com/starrocks/sql/optimizer/task/EnforceAndCostTask.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/DistributedEnvPlanWithCostTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/EnumeratePlanTest.java
… fe/fe-core/src/test/java/com/starrocks/sql/plan/ReplayFromDumpTest.java
… fe/fe-core/src/test/resources/sql/enumerate-plan/three-join.sql
… fe/fe-core/src/test/resources/sql/enumerate-plan/tpch-q10.sql
… fe/fe-core/src/test/resources/sql/enumerate-plan/tpch-q18.sql
… fe/fe-core/src/test/resources/sql/enumerate-plan/tpch-q21.sql
… fe/fe-core/src/test/resources/sql/enumerate-plan/tpch-q5.sql
… fe/fe-core/src/test/resources/sql/enumerate-plan/tpch-q7.sql
… fe/fe-core/src/test/resources/sql/enumerate-plan/tpch-q9.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q10.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q16.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q18.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q2.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q20.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q5.sql
… fe/fe-core/src/test/resources/sql/external/hive/tpch/q7.sql
… fe/fe-core/src/test/resources/sql/tpch-histogram-cost/q18.sql
… fe/fe-core/src/test/resources/sql/tpch-histogram-cost/q5.sql
… fe/fe-core/src/test/resources/sql/tpch-histogram-cost/q8.sql
… fe/fe-core/src/test/resources/sql/tpchcost/q18.sql
… fe/fe-core/src/test/resources/sql/tpchcost/q5.sql
… fe/fe-core/src/test/resources/sql/tpchcost/q8.sql