@wanpengfei-git 2022-10-30
starrocks_be_unittest succeeded.