@wanpengfei-git 2022-10-31
starrocks_be_unittest succeeded.