@packy92 2022-11-01
run starrocks_admit_test