@wanpengfei-git 2022-11-03
#### [FE PR Coverage Check]

:heart_eyes: **pass** : 99 / 102 (97.06%)




#### file detail

|   |path|covered_line|new_line|coverage|not_covered_line_detail|
|----|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/sql/optimizer/operator/physical/PhysicalDistributionOperator.java|0|1|00.00%|[54]|
|:large_blue_circle:|com/starrocks/sql/optimizer/cost/CostEstimate.java|1|2|50.00%|[72]|
|:large_blue_circle:|com/starrocks/sql/optimizer/cost/CostModel.java|35|36|97.22%|[234]|
|:large_blue_circle:|com/starrocks/sql/optimizer/cost/HashJoinCostModel.java|60|60|100.00%|[]|
|:large_blue_circle:|com/starrocks/sql/optimizer/task/EnforceAndCostTask.java|3|3|100.00%|[]|