@packy92 2022-11-04
run starrocks_admit_test